Add two useTorusWallet tests
