+++ __mocks__/@toruslabs/openlogin.js
@@ -1,0 +1,5 @@
+const OpenLogin = jest.createMockFromModule("@toruslabs/openlogin");
+
+OpenLogin.prototype.constructor = (...args) => {};
+
+export default OpenLogin;
